@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --extra-webpack-config webpack-dev.config.js",
+    "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
@@ -12,47 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.7",
-    "@angular/cdk": "^16.2.12",
-    "@angular/common": "^17.0.7",
-    "@angular/compiler": "^17.0.7",
-    "@angular/core": "^17.0.7",
-    "@angular/forms": "^17.0.7",
-    "@angular/material": "^16.2.12",
-    "@angular/platform-browser": "^17.0.7",
-    "@angular/platform-browser-dynamic": "^17.0.7",
-    "@angular/router": "^17.0.7",
+    "@angular/animations": "^20.3.7",
+    "@angular/cdk": "^20.2.10",
+    "@angular/common": "^20.3.7",
+    "@angular/compiler": "^20.3.7",
+    "@angular/core": "^20.3.7",
+    "@angular/forms": "^20.3.7",
+    "@angular/material": "^20.2.10",
+    "@angular/platform-browser": "^20.3.7",
+    "@angular/platform-browser-dynamic": "^20.3.7",
+    "@angular/router": "^20.3.7",
     "@mdi/angular-material": "^7.2.96",
-    "@ngneat/hot-toast": "^6.0.2",
-    "@ngneat/overview": "5.1.1",
+    "@ngneat/hot-toast": "^7.0.0",
+    "@ngneat/overview": "7.0.0",
     "@ngneat/transloco": "^6.0.4",
-    "angular-calendar": "^0.31.0",
-    "date-fns": "^2.29.3",
-    "dayjs": "^1.11.10",
+    "angular-calendar": "^0.32.0",
+    "date-fns": "^4.1.0",
+    "dayjs": "^1.11.18",
     "dragon": "^0.3.0",
-    "ng-for-track-by-property": "^16.0.1",
-    "ngx-drag-drop": "^17.0.0",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "ng-for-track-by-property": "^20.0.1",
+    "ngx-drag-drop": "^20.0.1",
+    "rxjs": "~7.8.2",
+    "tailwindcss": "^3.4.18",
+    "tslib": "^2.8.1",
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.7",
-    "@angular/cli": "~17.0.7",
-    "@angular/compiler-cli": "^17.0.7",
-    "@angular/localize": "^17.0.7",
+    "@angular-devkit/build-angular": "^20.3.7",
+    "@angular-devkit/core": "^20.3.7",
+    "@angular/build": "^20.3.7",
+    "@angular/cli": "^20.3.7",
+    "@angular/compiler-cli": "^20.3.7",
+    "@angular/localize": "^20.3.7",
     "@ngneat/transloco-keys-manager": "^3.8.0",
-    "@types/jasmine": "~4.3.0",
-    "autoprefixer": "^10.4.16",
-    "jasmine-core": "~4.6.0",
-    "karma": "~6.4.0",
+    "@types/jasmine": "~5.1.12",
+    "autoprefixer": "^10.4.21",
+    "jasmine-core": "~5.12.0",
+    "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "ngx-build-plus": "^17.0.0",
-    "postcss": "^8.4.32",
-    "tailwindcss": "^3.3.6",
-    "typescript": "~5.2.2"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "postcss": "^8.5.6",
+    "typescript": "~5.9.3"
   }
 }
