@@ -1,19 +1,20 @@
 import { Component, TemplateRef, ViewChild } from '@angular/core';
 import { HotToastService } from '@ngneat/hot-toast';
 import { catchError, of, tap, throwError } from 'rxjs';
-import { AppConfigService } from 'src/app/appconfig.service';
-import { GrocyService } from 'src/app/grocy.service';
+import { AppConfigService } from '../../appconfig.service';
+import { GrocyService } from '../../grocy.service';
 
 @Component({
-  selector: 'app-grocy-config',
-  templateUrl: './grocy-config.component.html',
-  styleUrl: './grocy-config.component.scss',
+    selector: 'app-grocy-config',
+    templateUrl: './grocy-config.component.html',
+    styleUrl: './grocy-config.component.scss',
+    standalone: false
 })
 export class GrocyConfigComponent {
   constructor(
     private appConfigService: AppConfigService,
     private grocyService: GrocyService
-  ) {}
+  ) { }
 
   @ViewChild('successTemplate') successTemplate: TemplateRef<any> | undefined;
   @ViewChild('errorTemplate') errorTemplate: TemplateRef<any> | undefined;
