@@ -1,9 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-made-by-banner',
-  templateUrl: './made-by-banner.component.html',
-  styleUrl: './made-by-banner.component.scss'
+    selector: 'app-made-by-banner',
+    templateUrl: './made-by-banner.component.html',
+    styleUrl: './made-by-banner.component.scss',
+    standalone: false
 })
 export class MadeByBannerComponent {
 
