@@ -1,7 +1,8 @@
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-  name: 'base64'
+    name: 'base64',
+    standalone: false
 })
 export class Base64Pipe implements PipeTransform {
 
